alternative CLI-based model download
